Hecha la pagina de notas completa, falta agregar estilos a las notas y el delete y edit

+++ app/src/main/java/es/fpsumma/dam2/utilidades/model/Asignatura.kt
@@ -19,4 +19,4 @@ data class Asignatura(
      val trimestre: String? = null,
 
     @ColumnInfo(name = "nota")
-     val nota: Float? = null)
+     val nota: Double? = null)
